split up classes into sub-modules
